Move modules into the backend folder
