+++ src/backend/modules/encryption/credentials.js
