+++ src/backend/modules/encryption/keytar.js
