+++ src/backend/modules/encryption/salt.js
